fix: panic safe logger initialization
